python_api/requirements.txt
tensorflow==2.15.0
Flask==3.0.2
Flask-RESTful==0.3.10
pillow==10.2.0
numpy==1.26.3